fix/mypage modules
