+++ .vscode/settings.json
@@ -10,6 +10,7 @@
     "gardensize",
     "Gmarket",
     "hookform",
+    "infoitem",
     "kakao",
     "MMKV",
     "mobily",
