+++ src/hooks/get-account-info/index.ts
@@ -1,0 +1,1 @@
+export * from './get-account-info-query.hook';
