+++ src/screens/primary/primary-my-page/module/primary-mypage-content/primary-mypage-greeting/index.ts
@@ -1,0 +1,1 @@
+export * from './primary-mypage-greeting.module';
