+++ src/screens/primary/primary-my-page/module/primary-mypage-content/primary-mypage-header/primary-mypage-header.module.tsx
@@ -8,7 +8,7 @@ import {
 import {
   PrimaryNavigatorParamLists,
   PrimaryNavigatorProps,
-} from '../../../primary.navigator';
+} from '../../../../primary.navigator';
 
 import { ModalHeader } from '@/layouts/modal-header/modal-header.layout';
