+++ src/screens/primary/primary-my-page/module/primary-mypage-content/primary-mypage-infoitem/index.ts
@@ -1,0 +1,1 @@
+export * from './primary-mypage-infoitem.module';
